% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{discover}
\alias{discover}
\title{Discover OpenID Connect Endpoints}
\usage{
discover(auth_server)
}
\arguments{
\item{auth_server}{the server}
}
\value{
An \code{\link{oauth_endpoint}} with the discovered endpoints.
}
\description{
Performs OpenID Connect discovery on an ID Provider.
}
\examples{
\dontrun{
discover("https://auth.molgenis.org")
discover("https://accounts.google.com")
}
}
