% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisAuth-package.R
\docType{package}
\name{MolgenisAuth-package}
\alias{MolgenisAuth}
\alias{MolgenisAuth-package}
\title{MolgenisAuth: Authentication Utils}
\description{
Helps the user authenticate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-auth}
  \item \url{https://molgenis.github.io/molgenis-r-auth}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-auth/issues}
}

}
\author{
\strong{Maintainer}: FDL Kelpin \email{f.kelpin@umcg.nl} (\href{https://orcid.org/0000-0003-1527-5130}{ORCID})

}
\keyword{internal}
