% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisAuth-package.R
\docType{package}
\name{MolgenisAuth-package}
\alias{MolgenisAuth}
\alias{MolgenisAuth-package}
\title{MolgenisAuth: 'OpenID Connect' Discovery and Authentication}
\description{
Discover 'OpenID Connect' endpoints and authenticate
    using device flow. Used by 'MOLGENIS' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-auth/}
  \item \url{https://molgenis.github.io/molgenis-r-auth/}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-auth/issues/}
}

}
\author{
\strong{Maintainer}: Sido Haakma \email{s.haakma@rug.nl} (\href{https://orcid.org/0000-0003-1278-9144}{ORCID})

Authors:
\itemize{
  \item Fleur Kelpin \email{fleur@kelpin.nl} (\href{https://orcid.org/0000-0003-1527-5130}{ORCID})
  \item Tommy de Boer \email{t.de.boer01@umcg.nl} (\href{https://orcid.org/0000-0002-4492-7225}{ORCID})
}

Other contributors:
\itemize{
  \item MOLGENIS org \email{molgenis-support@umcg.nl} [copyright holder, funder]
}

}
\keyword{internal}
